OUB-2241 Fix returning HTML on accident
Note from Jonathan: The new scrollparent module now uses the fallback of `document.scrollingElement` when no nested scroll area is found... this value used to be the body element... our check for an undefined scrollingElement wasn't correct. We need to check the return value of scrollparent for scrollHeight > clientHeight to make sure we return the window object correctly. (putting scroll listeners on html/body elements that do not have scrollHeight will not get scroll events)